lpha register. May have errors(see description)
Stuffs Added:
- Components needed to make registry
- Added route to registry(/register)

Possible error:
- In Inrupt's demo app when the register is successful the app redirects
you to the RegisterSucess page which waits 15 seconds to redirect you to
the login. We are not following this steps,out app redirects directly to
the login. We have have done this because at the moment we dont have the
Layouts components. In the future, if our app has problems with
register, we should follow Inrupt's steps(using useHistory hook).